Add files via upload
As a Data scientist at the Forest bank, and the bank has been facing a number of fraudulent transactions. Fradulent transaction are those which are not done by the user itself but by a hacker or by a card thief or by doing a fishing attack. The bank has a huge dataset of it's credit card users and the bank has also done the job of marking these transactions as fraudulent or not. Being banks' Data science expert I have been given the task to first analyse this data and generate insights and then predict if a transaction on a credit card is going to be fraud or not. I have used various classification algorithms like logistic regression, KNN, Decision tree, Random Forest, SVM to make predictions.